Move isRealJSONParse to its own file
